ocr-service/requirements.txt
# OCR Service Dependencies
# RapidOCR - 高精度中文OCR引擎
rapidocr-onnxruntime>=1.3.0

# RapidTable - 表格识别
rapid-table>=0.2.0

# FastAPI - Web框架
fastapi>=0.104.0
uvicorn>=0.24.0

# PDF处理
pdf2image>=1.16.0
Pillow>=10.0.0

# 其他工具
python-multipart>=0.0.6

# ===== RAG 向量检索相关 =====
# ChromaDB - 嵌入式向量数据库
chromadb>=0.4.0

# Sentence Transformers - 中文嵌入模型
sentence-transformers>=2.2.0

# 可选: 使用更轻量的中文模型
# text2vec-base-chinese 通过 sentence-transformers 加载

